requirements.txt
black==24.4.2
Django==4.2.2
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
dotenv==0.9.9
